Downloads your specific build of rustlib.
